web/pyproject.toml
[tool.poetry]
name = "wmanager"
version = "0.1.0"
description = ""
authors = ["MattheusBy <mattaudia6@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.8.2"
django = "^4.1.6"
djangorestframework = "^3.14.0"
psycopg2-binary = "^2.9.5"
djoser = "^2.1.0"
requests = "^2.28.2"
setuptools = "^67.2.0"
celery = "^5.2.7"
django-celery-results = "^2.4.0"
django-celery-beat = "^2.4.0"
django-redis = "^5.2.0"
redis-server = "^6.0.9"
flake8 = "^6.0.0"
gunicorn = "^20.1.0"
dj-database-url = "^1.2.0"



[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
